Refactored Visitors Logs
