+++ app/(security)/VisitorsLogs.tsx
@@ -1,29 +1,9 @@
 import React, { useEffect, useState } from 'react';
-import { View, Text, StyleSheet, TouchableOpacity, TextInput, ActivityIndicator, FlatList, Modal, ScrollView, SafeAreaView } from 'react-native';
-import { supabase } from '../lib/supabaseClient';
+import { View, Text, TouchableOpacity, TextInput, ActivityIndicator, FlatList, Modal, ScrollView, SafeAreaView } from 'react-native';
 import { Ionicons } from '@expo/vector-icons';
 import Logo from '../../components/logo';
-
-interface Visitor {
-  id: number;
-  name: string;
-  time_of_visit: string;
-  time_out?: string;
-  visit_id: string;
-  purpose_of_visit?: string;
-  phone_number?: string;
-  card_type?: string;
-  id_number?: string;
-  visit_count?: number;
-  expiration?: string; // Added field for expiration
-  visitors?: {
-    id: number;
-    name: string;
-    phone_number?: string;
-    card_type?: string;
-    id_number?: string;
-  };
-}
+import { Visitor, fetchVisitors, updateVisitorTimeOut, formatDate } from './api/notification-service/visitorsApi';
+import styles from '../../assets/VisitorsLogsStyles';
 
 const VisitorsLogs: React.FC = () => {
   const [visitors, setVisitors] = useState<Visitor[]>([]);
@@ -33,77 +13,37 @@ const VisitorsLogs: React.FC = () => {
   const [showVisitorModal, setShowVisitorModal] = useState(false);
   const [activeTab, setActiveTab] = useState('ongoing'); // 'ongoing' or 'completed'
   
-  // Date picker states
   const [showDateModal, setShowDateModal] = useState(false);
   const today = new Date();
   const [selectedDate, setSelectedDate] = useState(today);
   const [formattedDate, setFormattedDate] = useState(formatDate(today));
 
-  // Years for date picker
   const years = Array.from({ length: 10 }, (_, i) => today.getFullYear() - 5 + i);
   const months = [
     'January', 'February', 'March', 'April', 'May', 'June',
     'July', 'August', 'September', 'October', 'November', 'December'
   ];
- 
+        
   const getDaysInMonth = (month: number, year: number) => {
     return new Date(year, month, 0).getDate();
   };
-
+        
   useEffect(() => {
-    fetchVisitors();
+    loadVisitors();
   }, [formattedDate, activeTab]);
-
-  // Format date for display
-  function formatDate(date: Date): string {
-    const months = ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'];
-    const month = months[date.getMonth()];
-    const day = date.getDate();
-    const year = date.getFullYear();
-    return `${month} ${day}, ${year}`;
-  }
-
-  // Format time for display
-  function formatDateTime(dateString: string): string {
-    const date = new Date(dateString);
-    const months = ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'];
-    const month = months[date.getMonth()];
-    const day = date.getDate();
-    const year = date.getFullYear();
-    
-    let hours = date.getHours();
-    const minutes = date.getMinutes();
-    const ampm = hours >= 12 ? 'PM' : 'AM';
-    hours = hours % 12;
-    hours = hours ? hours : 12; // the hour '0' should be '12'
-    const minutesStr = minutes < 10 ? '0' + minutes : minutes;
-    
-    return `${month} ${day}, ${year} ${hours}:${minutesStr} ${ampm}`;
-  }
-
-  // Updated function to check if a visitor pass is expired
-  const isVisitorExpired = (visitor: Visitor): boolean => {
-    // If the visitor has an expiration field, use it
-    if (visitor.expiration) {
-      const expirationDate = new Date(visitor.expiration);
-      const now = new Date();
-      return now > expirationDate;
+        
+  const loadVisitors = async () => {
+    try {
+      setLoading(true);
+      const visitorData = await fetchVisitors(selectedDate, activeTab);
+      setVisitors(visitorData);
+    } catch (error) {
+      console.error('Error loading visitors:', error);
+    } finally {
+      setLoading(false);
     }
-    
-    // If no expiration date is available, fall back to the 10 PM same day logic
-    const visitDate = new Date(visitor.time_of_visit);
-    
-    // Set expiration to 10 PM (22:00) on the same day
-    const expirationDate = new Date(visitDate);
-    expirationDate.setHours(22, 0, 0, 0);
-    
-    // Adjust for timezone if needed
-    const expirationPht = new Date(expirationDate.getTime() - (8 * 60 * 60 * 1000));
-    
-    const now = new Date();
-    return now > expirationPht;
   };
-
+     
   const handleDateSelection = (year: number, month: number, day: number) => {
     const newDate = new Date(year, month - 1, day);
     setSelectedDate(newDate);
@@ -111,105 +51,18 @@ const VisitorsLogs: React.FC = () => {
     setShowDateModal(false);
   };
 
-  const fetchVisitors = async () => {
-    try {
-      setLoading(true);
-
-      // Get start and end of selected date
-      const startOfDay = new Date(new Date(selectedDate).setHours(0, 0, 0, 0)).toISOString();
-      const endOfDay = new Date(new Date(selectedDate).setHours(23, 59, 59, 999)).toISOString();
-      
-      // Modified query to fetch more visitor details including expiration
-      let query = supabase
-        .from('visits')
-        .select(`
-          id,
-          time_of_visit,
-          time_out,
-          visit_id,
-          purpose_of_visit,
-          expiration,
-          visitors(
-            id, 
-            name, 
-            phone_number, 
-            card_type, 
-            id_number
-          )
-        `)
-        .gte('created_at', startOfDay)
-        .lte('created_at', endOfDay);
-      
-      // Filter based on active tab
-      if (activeTab === 'ongoing') {
-        query = query.is('time_out', null);
-      } else {
-        query = query.not('time_out', 'is', null);
-      }
-      
-      const { data, error } = await query.order('time_of_visit', { ascending: false });
-      
-      if (error) throw error;
-
-      // Transform data with comprehensive visitor details and filter out expired visitors
-      const formattedData = data
-        .filter(item => item.visitors !== null)
-        .map(item => ({
-          id: item.id,
-          name: item.visitors?.name || 'Unknown Visitor',
-          time_of_visit: item.time_of_visit,
-          formatted_time_of_visit: formatDateTime(item.time_of_visit),
-          time_out: item.time_out,
-          formatted_time_out: item.time_out ? formatDateTime(item.time_out) : undefined,
-          visit_id: item.visit_id,
-          purpose_of_visit: item.purpose_of_visit || '',
-          phone_number: item.visitors?.phone_number || '',
-          card_type: item.visitors?.card_type || '',
-          id_number: item.visitors?.id_number || '',
-          visit_count: 3, // This would ideally be dynamically fetched
-          expiration: item.expiration
-        }))
-        .filter(visitor => {
-          // If it's an ongoing visit, don't show expired passes
-          if (activeTab === 'ongoing') {
-            return !isVisitorExpired(visitor);
-          }
-          // For completed visits, only show those that have been properly checked out
-          // and don't include expired but unchecked visitors
-          return visitor.time_out;
-        })
-        .map(visitor => ({
-          ...visitor,
-          time_of_visit: visitor.formatted_time_of_visit,
-          time_out: visitor.formatted_time_out
-        }));
-            
-      setVisitors(formattedData);
-    } catch (error) {
-      console.error('Error fetching visitors:', error);
-    } finally {
-      setLoading(false);
-    }
-  };
-
   const handleTimeOut = async (id: number) => {
     try {
-      const now = new Date().toISOString();
-      const { error } = await supabase
-        .from('visits')
-        .update({ time_out: now })
-        .eq('id', id);
-      
-      if (error) throw error;
+      await updateVisitorTimeOut(id);
       
       // Close the modal after time out
       setShowVisitorModal(false);
       // Refresh the list
-      fetchVisitors();
+      loadVisitors();
       // Switch to completed tab to show the logged out visitor
       setActiveTab('completed');
     } catch (error) {
-      console.error('Error updating time out:', error);
+      console.error('Error handling time out:', error);
     }
   };
 
@@ -244,7 +97,6 @@ const VisitorsLogs: React.FC = () => {
 
   return (
     <SafeAreaView style={styles.container}>
-      {/* Search moved to top */}
       <View style={styles.searchContainer}>
         <TextInput
           style={styles.searchInput}
@@ -257,7 +109,6 @@ const VisitorsLogs: React.FC = () => {
         </TouchableOpacity>
       </View>
       
-      {/* Tab Navigation */}
       <View style={styles.tabContainer}>
         <TouchableOpacity 
           style={[
@@ -286,10 +137,8 @@ const VisitorsLogs: React.FC = () => {
         </TouchableOpacity>
       </View>
       
-      {/* Date picker - Only visible in completed tab */}
       {activeTab === 'completed' && (
         <View style={styles.datePickerContainer}>
-          
           <TouchableOpacity
             style={styles.datePickerButton}
             onPress={() => setShowDateModal(true)}
@@ -297,7 +146,6 @@ const VisitorsLogs: React.FC = () => {
             <Text style={styles.dateText}>{formattedDate}</Text>
             <Ionicons name="calendar-outline" size={20} color="#252525" />
           </TouchableOpacity>
-          
         </View>
       )}
       
@@ -329,8 +177,7 @@ const VisitorsLogs: React.FC = () => {
         />
       )}
       
-      
-      {/* Date Picker Modal */}
+      {/* Date Selection Modal */}
       <Modal
         visible={showDateModal}
         transparent={true}
@@ -530,387 +377,9 @@ const VisitorsLogs: React.FC = () => {
         </View>
       </Modal>
 
-      <View style={styles.footer}>
-        
-      </View>
+      <View style={styles.footer}></View>
     </SafeAreaView>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#f5f5f5',
-  },
-  searchContainer: {
-    flexDirection: 'row',
-    margin: 16,
-    borderWidth: 1,
-    borderColor: '#e0e0e0',
-    borderRadius: 24,
-    backgroundColor: 'white',
-  },
-  searchInput: {
-    flex: 1,
-    padding: 10,
-    paddingLeft: 16,
-  },
-  searchButton: {
-    padding: 12,
-    justifyContent: 'center',
-    alignItems: 'center',
-    width: 48,
-  },
-  datePickerContainer: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    paddingHorizontal: 16,
-    paddingVertical: 10,
-  },
-  datePickerButton: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    padding: 8,
-    left: 250
-  },
-  dateText: {
-    marginRight: 10,
-    fontSize: 14,
-    fontWeight: '500',
-  },
-
-  // Tab Navigation Styles
-  tabContainer: {
-    flexDirection: 'row',
-    marginHorizontal: 16,
-    marginBottom: 16,
-    borderRadius: 8,
-    overflow: 'hidden',
-    borderWidth: 1,
-    borderColor: '#003566',
-  },
-  tabButton: {
-    flex: 1,
-    paddingVertical: 10,
-    alignItems: 'center',
-    backgroundColor: '#fff',
-  },
-  activeTabButton: {
-    backgroundColor: '#003566',
-  },
-  tabButtonText: {
-    fontWeight: '500',
-    color: '#000000',
-  },
-  activeTabButtonText: {
-    color: '#fff',
-  },
-  tableHeader: {
-    flexDirection: 'row',
-    paddingVertical: 10,
-    paddingHorizontal: 16,
-    backgroundColor: '#f0f0f0',
-    borderBottomWidth: 1,
-    borderBottomColor: '#e0e0e0',
-  },
-  headerText: {
-    fontWeight: '500',
-    color: '#555',
-  },
-  nameColumn: {
-    flex: 3,
-  },
-  timeColumn: {
-    flex: 4,
-    textAlign: 'center',
-  },
-  actionColumn: {
-    flex: 2,
-    textAlign: 'center',
-  },
-  listContainer: {
-    paddingBottom: 80,
-  },
-  visitorRow: {
-    flexDirection: 'row',
-    padding: 12,
-    backgroundColor: 'white',
-    borderBottomWidth: 1,
-    borderBottomColor: '#eeeeee',
-    alignItems: 'center',
-  },
-  visitorInfo: {
-    flex: 3,
-  },
-  visitorName: {
-    fontWeight: '500',
-    fontSize: 15,
-  },
-  timeInfo: {
-    flex: 4,
-    alignItems: 'center',
-  },
-  timeText: {
-    fontSize: 13,
-    textAlign: 'center',
-  },
-  timeInText: {
-    color: '#4CD964', // Green color for time in
-  },
-  timeOutText: {
-    color: '#FF3B30', // Red color for time out
-  },
-  actionContainer: {
-    flex: 2,
-    alignItems: 'center',
-  },
-  actionButton: {
-    backgroundColor: '#003566',
-    paddingVertical: 6,
-    paddingHorizontal: 12,
-    borderRadius: 4,
-  },
-  timeOutButton: {
-    backgroundColor: '#D9534F',
-  },
-  buttonText: {
-    color: 'white',
-    fontSize: 12,
-    fontWeight: '500',
-  },
-  loadingContainer: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  emptyContainer: {
-    padding: 20,
-    alignItems: 'center',
-  },
-  emptyText: {
-    color: '#252525',
-  },
-  footer: {
-    position: 'absolute',
-    bottom: 20,
-    width: '100%',
-    alignItems: 'center',
-  },
-
-
-  // Date Picker Modal Styles
-  modalContainer: {
-    flex: 1,
-    justifyContent: 'flex-end',
-    backgroundColor: 'rgba(0, 0, 0, 0.5)',
-  },
-  modalContent: {
-    backgroundColor: '#fff',
-    borderTopLeftRadius: 20,
-    borderTopRightRadius: 20,
-    padding: 20,
-    maxHeight: '70%',
-  },
-  modalHeader: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    marginBottom: 20,
-  },
-  modalTitle: {
-    fontSize: 18,
-    fontWeight: 'bold',
-  },
-  datePickerContainer2: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    marginBottom: 20,
-    height: 200,
-  },
-  pickerColumn: {
-    flex: 1,
-    marginHorizontal: 5,
-  },
-  pickerItem: {
-    padding: 10,
-    alignItems: 'center',
-    justifyContent: 'center',
-    borderRadius: 8,
-  },
-  selectedPickerItem: {
-    backgroundColor: '#e6f7ff',
-    borderWidth: 1,
-    borderColor: '#1890ff',
-  },
-  pickerText: {
-    fontSize: 16,
-  },
-  confirmButton: {
-    backgroundColor: '#000',
-    borderRadius: 8,
-    height: 50,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  confirmButtonText: {
-    color: '#fff',
-    fontSize: 16,
-    fontWeight: 'bold',
-  },
-  
-  // Visitor Details Modal Styles
-  modalOverlay: {
-    flex: 1,
-    backgroundColor: 'rgba(0, 0, 0, 0.5)',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  visitorModalContent: {
-    backgroundColor: '#fff',
-    borderRadius: 12,
-    width: '85%',
-    maxHeight: '80%',
-    paddingVertical: 20,
-    paddingHorizontal: 16,
-  },
-  closeButton: {
-    alignSelf: 'flex-end',
-    paddingVertical: 2,
-    paddingHorizontal: 2,
-    position: 'absolute',
-    top: 10,
-    right: 10,
-    zIndex: 10,
-  },
-  visitorDetailsContainer: {
-    width: '100%',
-  },
-  logoContainer: {
-    alignItems: 'center',
-    marginBottom: 16,
-  },
-  logoCircle: {
-    justifyContent: 'center',
-    alignItems: 'center',
-    marginBottom: 8,
-  },
-  universityInitials: {
-    color: '#fff',
-    fontSize: 18,
-    fontWeight: 'bold',
-  },
-  universityName: {
-    fontSize: 16,
-    fontWeight: '500',
-    textAlign: 'center',
-  },
-  visitorDetailsHeader: {
-    alignItems: 'center',
-    marginBottom: 20,
-  },
-  visitorDetailsName: {
-    fontSize: 22,
-    fontWeight: 'bold',
-    textAlign: 'center',
-    marginBottom: 11,
-  },
-  visitorId: {
-    fontSize: 14,
-    color: '#252525',
-    textAlign: 'center',
-  },
-  detailsContent: {
-    width: '100%',
-  },
-  detailsSection: {
-    flexDirection: 'row',
-    width: '100%',
-    paddingVertical: 8,
-    borderBottomWidth: 1,
-    borderBottomColor: '#eeeeee',
-  },
-  detailsLabel: {
-    flex: 1.5,
-    fontSize: 14,
-    color: '#555',
-  },
-  detailsValue: {
-    flex: 2,
-    fontSize: 14,
-    fontWeight: '500',
-  },
-  visitCountContainer: {
-    flex: 2,
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  visitCountValue: {
-    fontSize: 14,
-    fontWeight: '500',
-    marginRight: 10,
-  },
-  viewLogLink: {
-    color: '#4682B4',
-    textDecorationLine: 'underline',
-  },
-  timeSection: {
-    width: '100%',
-    marginTop: 16,
-  },
-  timeRow: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    paddingVertical: 6,
-  },
-  timeLabel: {
-    fontSize: 14,
-    color: '#555',
-  },
-  timeInValue: {
-    fontSize: 14,
-    fontWeight: '500',
-    color: '#4CD964',
-  },
-  timeOutValue: {
-    fontSize: 14,
-    fontWeight: '500',
-    color: '#FF3B30',
-  },
-  // Time Out button in modal
-  timeOutModalButton: {
-    backgroundColor: '#D9534F',
-    padding: 12,
-    borderRadius: 8,
-    marginTop: 20,
-    alignItems: 'center',
-  },
-  timeOutModalButtonText: {
-    color: 'white',
-    fontWeight: '600',
-    fontSize: 16,
-  },
-  // Additional styles for better UI
-  badgeContainer: {
-    borderRadius: 12,
-    paddingVertical: 2,
-    paddingHorizontal: 8,
-    marginLeft: 8,
-  },
-  activeBadge: {
-    backgroundColor: '#E1F5FE',
-  },
-  activeBadgeText: {
-    color: '#0288D1',
-    fontSize: 12,
-  },
-  expiredBadge: {
-    backgroundColor: '#FFEBEE',
-  },
-  expiredBadgeText: {
-    color: '#D32F2F',
-    fontSize: 12,
-  }
-});
 
 export default VisitorsLogs;
